KVM: PPC: Book3S HV: Add transactional memory support
This adds saving of the transactional memory (TM) checkpointed state on guest
entry and exit.  We only do this if we see that the guest has an active
transaction.

It also adds emulation of the TM state changes when delivering IRQs into the
guest.  According to the architecture, if we are transactional when an IRQ
occurs the TM state is changed to suspended, otherwise it's left unchanged.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Paul Mackerras <paulus@samba.org>
(cherry picked from commit a4ed18a21acf54750cd5c394e3cd1f8491a18bcc)