KVM: PPC: reserve a capability and KVM device type for in-kernel VFIO
This reserves a capability number for upcoming support
of VFIO-IOMMU DMA operations in real mode.

This reserves a number for a new "SPAPR TCE IOMMU" KVM device
which is going to manage lifetime of a list of SPAPR TCE IOMMU objects.

This defines an attribute of the "SPAPR TCE IOMMU" KVM device
which is going to be used by the userspace to communicate the KVM device.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
(cherry picked from commit c8277f0894635041e3d00c9fa2320a99a7ac9b6a)