KVM: PPC: Book3S HV: Fix bugs in IRQFD and update to latest submitted version
This fixes some bugs in the IRQFD support:

* kvm_notify_acked_irq() is supposed to be called on end-of-interrupt,
  not when the interrupt is fetched from the interrupt controller.
* Generic code calling kvm_set_irq will set level to 0 or 1, whereas
  calls from kvm_vm_ioctl_irq_line() as a result of the KVM_IRQ_LINE
  ioctl will usually use the powerpc-specific values KVM_INTERRUPT_SET,
  KVM_INTERRUPT_UNSET and KVM_INTERRUPT_SET_LEVEL.  Thus we need to
  cope with both sets of values.

This also simplifies the code and updates comments to reflect the fact
that the `line_status' argument to xics_set_irq is never going to be
true because we don't have the KVM_CAP_IRQ_INJECT_STATUS capability.
This brings the code into line with the most recent upstream submission.

Signed-off-by: Paul Mackerras <paulus@samba.org>
(cherry picked from commit 17e64f276f2d526daf86309bc49924620c69bdc7)