KVM: PPC: Book3S HV: Rework the secondary inhibit code
As part of the support for split core on power8, we want to be able to
block splitting of the core while KVM VMs are active.

The logic to do that would be exactly the same as the code we currently
have for inhibiting onlining of secondaries.

Instead of adding an identical mechanism to block split core, rework the
secondary inhibit code to be a "HV KVM is active" check. We can then use
that in both the cpu hotplug code and the upcoming split core code.

Signed-off-by: Michael Ellerman <michael@ellerman.id.au>
(cherry picked from commit 21c9d5ff0b0322eddd180632a778630eb525d75e)

Conflicts:
	arch/powerpc/include/asm/kvm_ppc.h
	arch/powerpc/include/asm/smp.h
	arch/powerpc/kvm/book3s_hv_builtin.c

(cherry picked from commit 5d9140ddf7fd4faf2a8f37b73e23442b149d1214)