powerpc: Add threads_per_subcore
On power8 we have a new concept of a subcore. This is what happens when
you take a regular core and split it. A subcore is a grouping of two or
four SMT threads, as well as a handfull of SPRs which allows the subcore
to appear as if it were a core from the point of view of a guest.

Unlike threads_per_core which is fixed at boot, threads_per_subcore can
change while the system is running. Most code will not want to use
threads_per_subcore.

Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Michael Neuling <mikey@neuling.org>
(cherry picked from commit bf8856d3f05aace850101bc4b02a3d60a14a331d)
(cherry picked from commit 30d68d5691c24d4ec1e70e95093add281fe18e7f)