powerpc/powernv: Add opal check token call
Currently there is no way to generically check if an opal call exists or not
from the host kernel.

This adds an opal call opal_check_token() which tells you if the given token is
present in opal or not.  This call was recently added to sapphire.

Signed-off-by: Michael Neuling <mikey@neuling.org>
(cherry picked from commit 24f55a3819f0f7a9dd627d2a2e0a5f80712d4502)