powerpc/powernv: Disable split core when opal timebase resync call is not present
Currently if the timebase resync call is not present we will BUG_ON() when we
try to split the core.

This patch changes this so that we don't allow core splitting when the timebase
resync call is not present, hence avoiding the later crash.

Signed-off-by: Michael Neuling <mikey@neuling.org>
(cherry picked from commit c94519c5aaf4e66e0eba5740db876d207fecce3a)